OrangeRoute project initial release
OrangeRoute is a shuttle tracking and route management platform designed for universities and smart campuses. Built with PHP, CSS, and JavaScript, it enables real-time vehicle monitoring, admin control, and user-friendly dashboards. This initial version focuses on scalable architecture and intuitive design for future expansion. This initial version was made to solve the shuttle tracking problem of UIU students